mcp-server/pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "Universal MCP connector for heterogeneous communication protocols"
authors = ["Kyle Adriany <kyle@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic-settings = "^2.0.3"
tenacity = "^8.2.3"
aiohttp = "^3.9.0"
pyserial-asyncio = "^0.6"
gmqtt = "^0.6.12"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
sentence-transformers = "^2.2.2"
numpy = "^1.26.0"
scikit-learn = "^1.3.2"
gitpython = "^3.1.40"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
mypy = "^1.7.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 88
target-version = ['py311']